Fare breakup not displaying in bottom info view

+++ AERTRIP/Modules/Flight/FlightDetails/FareBreakupVC/FareBreakupVC.swift
@@ -501,8 +501,8 @@ class FareBreakupVC: UIViewController, UITableViewDelegate, UITableViewDataSourc
                 baseFareCell.titleLabelLeading.constant = 31
                     if taxAndFeesData.count > 0{
                         baseFareCell.titleLabel.text = (taxAndFeesData[indexPath.row-1]["name"] as? String ?? "")
-                        if (taxAndFeesData[indexPath.row-1]["value"] as? Int) != nil{
-                            let amount : Double = Double(taxAndFeesData[indexPath.row-1]["value"] as? Int ?? 0)
+                        if (taxAndFeesData[indexPath.row-1]["value"] as? Double) != nil{
+                            let amount : Double = Double(taxAndFeesData[indexPath.row-1]["value"] as? Double ?? 0)
                             
                             let price = displayPriceInFormat(price: amount, fromOption : "FareAmount")
                             baseFareCell.amountLable.attributedText = price
